refactor(server): split cloud command lease store
