+++ server/proliferate/server/cloud/commands/wake.py
@@ -9,7 +9,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from proliferate.background.config import RUNTIME_WAKE_QUEUE, RUNTIME_WAKE_TARGET_TASK
 from proliferate.constants.cloud import CloudCommandStatus
 from proliferate.db.store.background_outbox import enqueue_outbox_task
-from proliferate.db.store.cloud_sync import commands as commands_store
+from proliferate.db.store.cloud_sync import command_records
 from proliferate.db.store.cloud_sync import targets as targets_store
 from proliferate.server.cloud.commands.domain.target import target_requires_cloud_workspace
 from proliferate.server.cloud.live.service import publish_worker_control_after_commit
@@ -47,7 +47,7 @@ def is_terminal_command_status(status: str) -> bool:
 
 def _command_requires_managed_target_wake(
     target: targets_store.CloudTargetSnapshot,
-    command: commands_store.CloudCommandSnapshot,
+    command: command_records.CloudCommandSnapshot,
 ) -> bool:
     if is_terminal_command_status(command.status):
         return False
@@ -60,7 +60,7 @@ async def kick_off_command_wake_after_commit_if_required(
     db: AsyncSession,
     *,
     target: targets_store.CloudTargetSnapshot,
-    command: commands_store.CloudCommandSnapshot,
+    command: command_records.CloudCommandSnapshot,
 ) -> None:
     await publish_worker_control_after_commit(db, target_id=target.id, reason="command")
     if not _command_requires_managed_target_wake(target, command):
